Layers.ConvBase: * removed unused Update() implementation
Layers:
*added TrainableBase for potential future implementation (both conv and dense share same methods.)

Network:
*not automatically initialized
*iniitalize is now a public method

Optimizers.Decays:
*bug fixes

Utils:
*added benchmark class which measures net performence

+++ GNet/Optimizers/Decays/Multiplication.cs
@@ -17,7 +17,7 @@ namespace GNet.Optimizers.Decays
 
         public double Compute(double value, int iteration)
         {
-            return value * Pow(Multiplier, iteration / Interval);
+            return value * Pow(Multiplier, (double)iteration / Interval);
         }
     }
 }
